Test provided endpoints in config at startup
